frontend, engine: extract the god-file seams
Pure code motion, byte-verified against HEAD by the extraction passes
(only visibility keywords differ):

- app.rs (3613 -> 2863): the ~600-line editor-playtest disc-authoring
  cluster (cue/bin writing, system area, world/UI pack payloads,
  build metadata) moves to frontend/src/playtest_disc.rs with a
  9-item pub(crate) surface; its 6 tests move along.
- cli.rs (2926 -> 1962): three focused submodules -- cli/headless_log
  (CSV/telemetry loggers), cli/profile_report (guest-profile
  printers), cli/iso_inspect (ISO-9660 reader) -- each exposing only
  what cli.rs calls.
- game_app.rs (3129 -> 2916): the transition/full-screen rendering
  free functions move to psx-engine/src/transitions.rs; one
  pub(crate) entry point.

Verified on the combined tree: workspace check clean across all
targets, 114 frontend + 433 core + 126 gpu-render + 209 engine tests
green, MIPS guest builds, canonical alttp hashes bit-identical, and
make validate passes 2/2.
