+++ emu/crates/frontend/src/cli.rs
@@ -23,13 +23,13 @@
 
 #[cfg(feature = "editor")]
 use std::collections::HashSet;
-use std::io::{BufWriter, Write};
+use std::io::Write;
 use std::path::{Path, PathBuf};
 
 use clap::{Args, Parser, Subcommand};
 use emulator_core::{
     button, fast_boot_disc_with_hle, spu::SAMPLE_CYCLES, telemetry, warm_bios_for_disc_fast_boot,
-    Bus, ButtonState, Cpu, GteProfileSnapshot, DISC_FAST_BOOT_WARMUP_STEPS,
+    Bus, ButtonState, Cpu, DISC_FAST_BOOT_WARMUP_STEPS,
 };
 // `Gpu` is only constructed for the editor 3D preview dump.
 #[cfg(feature = "editor")]
@@ -47,25 +47,18 @@ use psxed_project::{NodeId, ProjectDocument};
 #[cfg(feature = "editor")]
 use psxed_ui::{ViewportCameraMode, ViewportCameraState};
 
-#[cfg(feature = "editor")]
-use crate::app::{build_embedded_playtest_disc, copy_project_disc};
 use crate::app::{bus_from_configured_bios, fast_boot_embedded_playtest_disc};
+#[cfg(feature = "editor")]
+use crate::playtest_disc::{build_embedded_playtest_disc, copy_project_disc};
 use crate::playtest_input::read_input_tape;
 
-const NTSC_CPU_CYCLES_PER_VBLANK: u64 = 33_868_800 / 60;
-const GUEST_RENDER_BREAKDOWN_STAGES: &[(u16, &str)] = &[
-    (telemetry::stage::SKY, "sky"),
-    (telemetry::stage::FAR_VISTA, "far vista"),
-    (telemetry::stage::ROOM, "room"),
-    (telemetry::stage::ENTITY_MARKERS, "markers"),
-    (telemetry::stage::IMAGE_PROPS, "image props"),
-    (telemetry::stage::MODEL_INSTANCES, "models"),
-    (telemetry::stage::PLAYER, "player"),
-    (telemetry::stage::EQUIPMENT, "equipment"),
-    (telemetry::stage::WORLD_FLUSH, "flush/sort"),
-    (telemetry::stage::OT_SUBMIT, "ot submit"),
-    (telemetry::stage::OT_WAIT, "ot wait"),
-];
+mod headless_log;
+mod iso_inspect;
+mod profile_report;
+
+use headless_log::{CounterLog, DisplayHashLog, GuestProfileLog};
+use iso_inspect::{contains_bytes, iso_root_entries, iso_volume_id};
+use profile_report::{print_gte_profile, print_guest_profile};
 
 /// Top-level argument parser. Passed to `clap::Parser::parse()`
 /// from `main.rs`.
@@ -1115,7 +1108,8 @@ fn write_wav_s16_stereo(
 
 #[cfg(feature = "editor")]
 fn cmd_build_editor_playtest_disc() -> Result<(), String> {
-    let disc_path = build_embedded_playtest_disc(crate::app::DEFAULT_EMBEDDED_PLAYTEST_VOLUME_ID)?;
+    let disc_path =
+        build_embedded_playtest_disc(crate::playtest_disc::DEFAULT_EMBEDDED_PLAYTEST_VOLUME_ID)?;
     println!("{}", disc_path.display());
     Ok(())
 }
@@ -1125,20 +1119,6 @@ fn cmd_build_project_disc(args: BuildProjectDiscArgs) -> Result<(), String> {
     let dest_cue = build_project_disc_path(&args.project)?;
     println!("{}", dest_cue.display());
     Ok(())
-}
-
-#[derive(Clone, Debug)]
-struct IsoRootEntry {
-    identifier: String,
-    extent_lba: u32,
-    size: u32,
-    flags: u8,
-}
-
-impl IsoRootEntry {
-    fn is_dir(&self) -> bool {
-        self.flags & 0x02 != 0
-    }
 }
 
 fn cmd_preburn_check(args: PreburnCheckArgs) -> Result<(), String> {
@@ -1229,94 +1209,6 @@ fn cmd_preburn_check(args: PreburnCheckArgs) -> Result<(), String> {
     Ok(())
 }
 
-fn iso_volume_id(disc: &Disc) -> Result<String, String> {
-    let pvd = disc
-        .read_sector_user(16)
-        .ok_or_else(|| "missing ISO9660 primary volume descriptor at LBA 16".to_string())?;
-    if pvd.len() < 72 || pvd[0] != 1 || &pvd[1..6] != b"CD001" || pvd[6] != 1 {
-        return Err("LBA 16 is not an ISO9660 primary volume descriptor".to_string());
-    }
-    Ok(String::from_utf8_lossy(&pvd[40..72]).trim().to_string())
-}
-
-fn iso_root_entries(disc: &Disc) -> Result<Vec<IsoRootEntry>, String> {
-    let pvd = disc
-        .read_sector_user(16)
-        .ok_or_else(|| "missing ISO9660 primary volume descriptor at LBA 16".to_string())?;
-    if pvd.len() < 190 || pvd[0] != 1 || &pvd[1..6] != b"CD001" || pvd[6] != 1 {
-        return Err("LBA 16 is not an ISO9660 primary volume descriptor".to_string());
-    }
-    let root =
-        parse_iso_dir_record(&pvd[156..]).ok_or_else(|| "bad root directory record".to_string())?;
-    if !root.is_dir() {
-        return Err("ISO9660 root record is not a directory".to_string());
-    }
-    let bytes = read_iso_extent(disc, root.extent_lba, root.size)?;
-    let mut entries = Vec::new();
-    let mut offset = 0usize;
-    while offset < bytes.len() {
-        let len = bytes[offset] as usize;
-        if len == 0 {
-            offset =
-                ((offset / psx_iso::SECTOR_USER_DATA_BYTES) + 1) * psx_iso::SECTOR_USER_DATA_BYTES;
-            continue;
-        }
-        let end = offset
-            .checked_add(len)
-            .filter(|end| *end <= bytes.len())
-            .ok_or_else(|| "bad ISO9660 directory record length".to_string())?;
-        let entry = parse_iso_dir_record(&bytes[offset..end])
-            .ok_or_else(|| "bad ISO9660 directory record".to_string())?;
-        if entry.identifier != "\u{0}" && entry.identifier != "\u{1}" {
-            entries.push(entry);
-        }
-        offset = end;
-    }
-    Ok(entries)
-}
-
-fn parse_iso_dir_record(record: &[u8]) -> Option<IsoRootEntry> {
-    let len = *record.first()? as usize;
-    if len < 34 || record.len() < len {
-        return None;
-    }
-    let ident_len = *record.get(32)? as usize;
-    let ident_start = 33usize;
-    let ident_end = ident_start.checked_add(ident_len)?;
-    if ident_end > len {
-        return None;
-    }
-    Some(IsoRootEntry {
-        extent_lba: u32::from_le_bytes(record.get(2..6)?.try_into().ok()?),
-        size: u32::from_le_bytes(record.get(10..14)?.try_into().ok()?),
-        flags: *record.get(25)?,
-        identifier: String::from_utf8(record[ident_start..ident_end].to_vec()).ok()?,
-    })
-}
-
-fn read_iso_extent(disc: &Disc, extent_lba: u32, size: u32) -> Result<Vec<u8>, String> {
-    let mut bytes = Vec::with_capacity(size as usize);
-    let mut remaining = size as usize;
-    let mut lba = extent_lba;
-    while remaining > 0 {
-        let sector = disc
-            .read_sector_user(lba)
-            .ok_or_else(|| format!("unreadable ISO9660 extent sector LBA {lba}"))?;
-        let take = remaining.min(sector.len());
-        bytes.extend_from_slice(&sector[..take]);
-        remaining -= take;
-        lba = lba.saturating_add(1);
-    }
-    Ok(bytes)
-}
-
-fn contains_bytes(haystack: &[u8], needle: &[u8]) -> bool {
-    !needle.is_empty()
-        && haystack
-            .windows(needle.len())
-            .any(|window| window == needle)
-}
-
 #[cfg(feature = "editor")]
 fn build_project_disc_path(project_path: &Path) -> Result<PathBuf, String> {
     let (project_root, project_file) = resolve_project_arg(project_path);
@@ -1335,7 +1227,7 @@ fn build_project_disc_path(project_path: &Path) -> Result<PathBuf, String> {
     )?;
     run_make(&repo_root, "build-editor-playtest", &[])?;
 
-    let volume_id = crate::app::project_disc_volume_id(&project.name);
+    let volume_id = crate::playtest_disc::project_disc_volume_id(&project.name);
     let source_cue = build_embedded_playtest_disc(&volume_id)?;
     let dest_cue = project_root.join("baked").join(format!(
         "{}.cue",
@@ -1802,552 +1694,6 @@ fn hash_vram(bus: &Bus) -> u64 {
     h.finish()
 }
 
-struct DisplayHashLog {
-    writer: Option<BufWriter<std::fs::File>>,
-    interval: u64,
-    checkpoint_kind: &'static str,
-}
-
-impl DisplayHashLog {
-    fn new(
-        path: Option<&Path>,
-        interval: u64,
-        checkpoint_kind: &'static str,
-    ) -> Result<Self, String> {
-        let Some(path) = path else {
-            return Ok(Self {
-                writer: None,
-                interval: interval.max(1),
-                checkpoint_kind,
-            });
-        };
-        if let Some(parent) = path
-            .parent()
-            .filter(|parent| !parent.as_os_str().is_empty())
-        {
-            std::fs::create_dir_all(parent)
-                .map_err(|e| format!("mkdir {}: {e}", parent.display()))?;
-        }
-        let file =
-            std::fs::File::create(path).map_err(|e| format!("create {}: {e}", path.display()))?;
-        let mut writer = BufWriter::new(file);
-        writeln!(
-            writer,
-            "checkpoint_kind,checkpoint_frame,guest_frame,visual_frame,cpu_tick,bus_cycles,display_hash,width,height,byte_len"
-        )
-        .map_err(|e| format!("write {}: {e}", path.display()))?;
-        Ok(Self {
-            writer: Some(writer),
-            interval: interval.max(1),
-            checkpoint_kind,
-        })
-    }
-
-    fn record(
-        &mut self,
-        checkpoint_frame: u64,
-        guest_frame: u64,
-        visual_frame: u64,
-        cpu_tick: u64,
-        bus_cycles: u64,
-        bus: &Bus,
-    ) -> Result<(), String> {
-        let Some(writer) = self.writer.as_mut() else {
-            return Ok(());
-        };
-        if checkpoint_frame % self.interval != 0 {
-            return Ok(());
-        }
-        let (hash, width, height, byte_len) = bus.gpu.display_hash();
-        writeln!(
-            writer,
-            "{},{checkpoint_frame},{guest_frame},{visual_frame},{cpu_tick},{bus_cycles},0x{hash:016x},{width},{height},{byte_len}",
-            self.checkpoint_kind
-        )
-        .map_err(|e| format!("write visual hash log: {e}"))
-    }
-
-    fn flush(&mut self) -> Result<(), String> {
-        if let Some(writer) = self.writer.as_mut() {
-            writer
-                .flush()
-                .map_err(|e| format!("flush visual hash log: {e}"))?;
-        }
-        Ok(())
-    }
-}
-
-/// Per-guest-frame log of frametime + the portal/streaming masks and camera
-/// position, so a streaming/visibility change can be measured frame-by-frame
-/// and a drawn-but-not-visible room pinned to an exact camera position.
-struct CounterLog {
-    writer: Option<BufWriter<std::fs::File>>,
-}
-
-impl CounterLog {
-    fn new(path: Option<&Path>) -> Result<Self, String> {
-        let Some(path) = path else {
-            return Ok(Self { writer: None });
-        };
-        if let Some(parent) = path
-            .parent()
-            .filter(|parent| !parent.as_os_str().is_empty())
-        {
-            std::fs::create_dir_all(parent)
-                .map_err(|e| format!("mkdir {}: {e}", parent.display()))?;
-        }
-        let file =
-            std::fs::File::create(path).map_err(|e| format!("create {}: {e}", path.display()))?;
-        let mut writer = BufWriter::new(file);
-        writeln!(
-            writer,
-            "guest_frame,cpu_tick,bus_cycles,resident_mask,active_mask,drawn_mask,visible_mask,cam_x_biased,cam_y_biased,cam_z_biased,current_room"
-        )
-        .map_err(|e| format!("write {}: {e}", path.display()))?;
-        Ok(Self {
-            writer: Some(writer),
-        })
-    }
-
-    fn record(
-        &mut self,
-        guest_frame: u64,
-        cpu_tick: u64,
-        bus_cycles: u64,
-        bus: &Bus,
-    ) -> Result<(), String> {
-        let Some(writer) = self.writer.as_mut() else {
-            return Ok(());
-        };
-        use telemetry::counter as c;
-        let g = |id| bus.telemetry.counter_latest_value(id);
-        writeln!(
-            writer,
-            "{guest_frame},{cpu_tick},{bus_cycles},{},{},{},{},{},{},{},{}",
-            g(c::ROOM_STREAM_RESIDENT_MASK_LO),
-            g(c::ROOM_ACTIVE_CHUNK_MASK_LO),
-            g(c::ROOM_DRAWN_CHUNK_MASK_LO),
-            g(c::PORTAL_VIS_VISIBLE_MASK_LO),
-            g(c::ROOM_CAMERA_GLOBAL_X_BIASED),
-            g(c::ROOM_CAMERA_GLOBAL_Y_BIASED),
-            g(c::ROOM_CAMERA_GLOBAL_Z_BIASED),
-            g(c::PORTAL_VIS_CURRENT_ROOM),
-        )
-        .map_err(|e| format!("write counter log: {e}"))
-    }
-
-    fn flush(&mut self) -> Result<(), String> {
-        if let Some(writer) = self.writer.as_mut() {
-            writer
-                .flush()
-                .map_err(|e| format!("flush counter log: {e}"))?;
-        }
-        Ok(())
-    }
-}
-
-const PROFILE_LOG_HEADER: &[&str] = &[
-    "guest_frame",
-    "end_cpu_tick",
-    "start_bus_cycles",
-    "end_bus_cycles",
-    "frame_cycles",
-    "frame_markers",
-    "fixed_update_task",
-    "visual_render_task",
-    "update",
-    "render",
-    "present",
-    "frame_clear",
-    "camera",
-    "portal_visibility",
-    "active_room_window",
-    "room_surface_cache",
-    "vram_upload",
-    "cd_room_chunk_load",
-    "room",
-    "room_visible_list",
-    "room_cell_select",
-    "room_project",
-    "room_depth_prep",
-    "room_surface_draw",
-    "sky",
-    "far_vista",
-    "image_props",
-    "box_props",
-    "box_prop_debris",
-    "box_prop_shards",
-    "image_cards",
-    "model_instances",
-    "model_bounds",
-    "model_draw",
-    "player",
-    "player_bounds",
-    "player_draw",
-    "equipment",
-    "textured_model_joints",
-    "textured_model_project",
-    "textured_model_faces",
-    "world_flush",
-    "ot_submit",
-    "ot_wait",
-    "sim_collision",
-    "sim_room_track",
-    "sim_residency",
-    "sim_pump",
-    "sim_solve",
-    "update_actor",
-    "update_window",
-    "cell_lookup",
-    "cell_depth",
-    "cell_collect",
-    "sim_ticks",
-    "visual_frames",
-    "visual_skipped_vblanks",
-    "visual_deadline_misses",
-    "visual_lateness_vblanks",
-    "cd_room_chunk_loads",
-    "cd_room_chunk_bytes",
-    "cd_room_chunk_sectors",
-    "cd_room_chunk_status",
-    "room_stream_requests",
-    "room_stream_misses",
-    "room_stream_prefetch_requests",
-    "room_stream_pending_loads",
-    "room_stream_evicts",
-    "room_stream_failed_loads",
-    "room_stream_resident_slots",
-    "room_stream_loading_mask",
-    "room_active_chunks",
-    "room_visible_cells",
-    "room_cells_drawn",
-    "room_cells_considered",
-    "room_cells_culled",
-    "room_cells_range_culled",
-    "room_vis_fallback_draws",
-    "room_surf_material",
-    "room_surf_projected",
-    "room_surf_screen",
-    "room_surf_kind",
-    "room_surf_backface",
-    "room_surf_lighting",
-    "room_surf_submit",
-    "room_surf_profiled",
-    "room_surf_screen_culled",
-    "room_surf_backface_culled",
-    "room_submit_hw_safe_test",
-    "room_submit_packet_fill",
-    "room_submit_primitive_push",
-    "room_submit_depth",
-    "room_submit_command",
-    "room_submit_fallback",
-    "room_surfaces_considered",
-    "room_projected_vertices",
-    "room_surf_whole_quads",
-    "room_surf_split_tris",
-    "tri_primitives",
-    "tri_primitive_remaining",
-    "world_commands",
-    "room_submit_primitive_overflows",
-    "model_instance_draws",
-    "player_projected_vertices",
-    "player_submitted_tris",
-    "textured_model_parts",
-    "textured_model_vertices",
-    "textured_model_faces_counter",
-    "textured_model_fast_submitted_tris",
-    "textured_model_split_tris",
-    "textured_model_packed_unclamped",
-    "textured_model_packed_clamped",
-    "textured_model_packed_general",
-    "textured_model_fallback_faces",
-    "resident_mask",
-    "active_mask",
-    "drawn_mask",
-    "visible_mask",
-    "missing_mask",
-    "camera_x_biased",
-    "camera_y_biased",
-    "camera_z_biased",
-    "player_view_yaw_q12",
-    "current_room",
-    "player_room",
-];
-
-struct ProfileFrame {
-    guest_frame: u64,
-    start_bus_cycles: u64,
-    events: Vec<telemetry::GuestTelemetryEvent>,
-}
-
-/// Per-completed-guest-frame log of runtime telemetry, used to identify which
-/// hot systems overlap on deadline-miss spikes in a recorded run.
-struct GuestProfileLog {
-    writer: Option<BufWriter<std::fs::File>>,
-    current_frame: Option<ProfileFrame>,
-}
-
-impl GuestProfileLog {
-    fn new(path: Option<&Path>) -> Result<Self, String> {
-        let Some(path) = path else {
-            return Ok(Self {
-                writer: None,
-                current_frame: None,
-            });
-        };
-        if let Some(parent) = path
-            .parent()
-            .filter(|parent| !parent.as_os_str().is_empty())
-        {
-            std::fs::create_dir_all(parent)
-                .map_err(|e| format!("mkdir {}: {e}", parent.display()))?;
-        }
-        let file =
-            std::fs::File::create(path).map_err(|e| format!("create {}: {e}", path.display()))?;
-        let mut writer = BufWriter::new(file);
-        writeln!(writer, "{}", PROFILE_LOG_HEADER.join(","))
-            .map_err(|e| format!("write {}: {e}", path.display()))?;
-        Ok(Self {
-            writer: Some(writer),
-            current_frame: None,
-        })
-    }
-
-    fn add_events(
-        &mut self,
-        events: &[telemetry::GuestTelemetryEvent],
-        end_cpu_tick: u64,
-        end_bus_cycles: u64,
-    ) -> Result<(), String> {
-        if self.writer.is_none() {
-            return Ok(());
-        }
-        for &event in events {
-            if matches!(event.kind, telemetry::GuestTelemetryKind::FrameBegin) {
-                self.flush_current(end_cpu_tick, event.cycles)?;
-                self.current_frame = Some(ProfileFrame {
-                    guest_frame: event.value as u64,
-                    start_bus_cycles: event.cycles,
-                    events: vec![event],
-                });
-            } else if let Some(frame) = self.current_frame.as_mut() {
-                frame.events.push(event);
-            }
-        }
-        if events.is_empty() {
-            self.flush_current(end_cpu_tick, end_bus_cycles)?;
-        }
-        Ok(())
-    }
-
-    fn finish(&mut self, end_cpu_tick: u64, end_bus_cycles: u64) -> Result<(), String> {
-        self.flush_current(end_cpu_tick, end_bus_cycles)
-    }
-
-    fn flush(&mut self) -> Result<(), String> {
-        if let Some(writer) = self.writer.as_mut() {
-            writer
-                .flush()
-                .map_err(|e| format!("flush profile log: {e}"))?;
-        }
-        Ok(())
-    }
-
-    fn flush_current(&mut self, end_cpu_tick: u64, end_bus_cycles: u64) -> Result<(), String> {
-        let Some(frame) = self.current_frame.take() else {
-            return Ok(());
-        };
-        self.write_frame(frame, end_cpu_tick, end_bus_cycles)
-    }
-
-    fn write_frame(
-        &mut self,
-        frame: ProfileFrame,
-        end_cpu_tick: u64,
-        end_bus_cycles: u64,
-    ) -> Result<(), String> {
-        let Some(writer) = self.writer.as_mut() else {
-            return Ok(());
-        };
-        let summary = telemetry::GuestTelemetrySummary::from_events(&frame.events);
-        let stage = |id: u16| {
-            summary
-                .stage_cycles
-                .get(id as usize)
-                .copied()
-                .unwrap_or_default()
-        };
-        let task = |id: u16| {
-            summary
-                .task_cycles
-                .get(id as usize)
-                .copied()
-                .unwrap_or_default()
-        };
-        let counter_latest = |id: u16| {
-            summary
-                .counter_latest_values
-                .get(id as usize)
-                .copied()
-                .unwrap_or_default()
-        };
-        let mut row = Vec::with_capacity(PROFILE_LOG_HEADER.len());
-        macro_rules! push {
-            ($value:expr) => {
-                row.push(($value).to_string())
-            };
-        }
-
-        use telemetry::{counter as c, stage as s, task as t};
-        push!(frame.guest_frame);
-        push!(end_cpu_tick);
-        push!(frame.start_bus_cycles);
-        push!(end_bus_cycles);
-        push!(end_bus_cycles.saturating_sub(frame.start_bus_cycles));
-        push!(summary.frames);
-        push!(task(t::FIXED_UPDATE));
-        push!(task(t::VISUAL_RENDER));
-        push!(stage(s::UPDATE));
-        push!(stage(s::RENDER));
-        push!(stage(s::PRESENT));
-        push!(stage(s::FRAME_CLEAR));
-        push!(stage(s::CAMERA));
-        push!(stage(s::PORTAL_VISIBILITY));
-        push!(stage(s::ACTIVE_ROOM_WINDOW));
-        push!(stage(s::ROOM_SURFACE_CACHE));
-        push!(stage(s::VRAM_UPLOAD));
-        push!(stage(s::CD_ROOM_CHUNK_LOAD));
-        push!(stage(s::ROOM));
-        push!(stage(s::ROOM_VISIBLE_LIST));
-        push!(stage(s::ROOM_CELL_SELECT));
-        push!(stage(s::ROOM_PROJECT));
-        push!(stage(s::ROOM_DEPTH_PREP));
-        push!(stage(s::ROOM_SURFACE_DRAW));
-        push!(stage(s::SKY));
-        push!(stage(s::FAR_VISTA));
-        push!(stage(s::IMAGE_PROPS));
-        push!(stage(s::BOX_PROPS));
-        push!(stage(s::BOX_PROP_DEBRIS));
-        push!(stage(s::BOX_PROP_SHARDS));
-        push!(stage(s::IMAGE_CARDS));
-        push!(stage(s::MODEL_INSTANCES));
-        push!(stage(s::MODEL_BOUNDS));
-        push!(stage(s::MODEL_DRAW));
-        push!(stage(s::PLAYER));
-        push!(stage(s::PLAYER_BOUNDS));
-        push!(stage(s::PLAYER_DRAW));
-        push!(stage(s::EQUIPMENT));
-        push!(stage(s::TEXTURED_MODEL_JOINTS));
-        push!(stage(s::TEXTURED_MODEL_PROJECT));
-        push!(stage(s::TEXTURED_MODEL_FACES));
-        push!(stage(s::WORLD_FLUSH));
-        push!(stage(s::OT_SUBMIT));
-        push!(stage(s::OT_WAIT));
-        push!(stage(s::SIM_COLLISION));
-        push!(stage(s::SIM_ROOM_TRACK));
-        push!(stage(s::SIM_RESIDENCY));
-        push!(stage(s::SIM_PUMP));
-        push!(stage(s::SIM_SOLVE));
-        push!(stage(s::UPDATE_ACTOR));
-        push!(stage(s::UPDATE_WINDOW));
-        push!(stage(s::CELL_LOOKUP));
-        push!(stage(s::CELL_DEPTH));
-        push!(stage(s::CELL_COLLECT));
-        push!(counter_total(&summary, c::SIM_TICKS));
-        push!(counter_total(&summary, c::VISUAL_FRAMES));
-        push!(counter_total(&summary, c::VISUAL_SKIPPED_VBLANKS));
-        push!(counter_total(&summary, c::VISUAL_DEADLINE_MISSES));
-        push!(counter_latest(c::VISUAL_MAX_LATENESS_VBLANKS));
-        push!(counter_total(&summary, c::CD_ROOM_CHUNK_LOADS));
-        push!(counter_total(&summary, c::CD_ROOM_CHUNK_BYTES));
-        push!(counter_total(&summary, c::CD_ROOM_CHUNK_SECTORS));
-        push!(counter_latest(c::CD_ROOM_CHUNK_STATUS));
-        push!(counter_total(&summary, c::ROOM_STREAM_REQUESTS));
-        push!(counter_total(&summary, c::ROOM_STREAM_MISSES));
-        push!(counter_total(&summary, c::ROOM_STREAM_PREFETCH_REQUESTS));
-        push!(counter_total(&summary, c::ROOM_STREAM_PENDING_LOADS));
-        push!(counter_total(&summary, c::ROOM_STREAM_EVICTIONS));
-        push!(counter_total(&summary, c::ROOM_STREAM_FAILED_LOADS));
-        push!(counter_latest(c::ROOM_STREAM_RESIDENT_SLOTS));
-        push!(counter_latest(c::ROOM_STREAM_LOADING_MASK_LO));
-        push!(counter_total(&summary, c::ROOM_ACTIVE_CHUNKS));
-        push!(counter_total(&summary, c::ROOM_VISIBLE_CELLS));
-        push!(counter_total(&summary, c::ROOM_CELLS_DRAWN));
-        push!(counter_total(&summary, c::ROOM_CELLS_CONSIDERED));
-        push!(counter_total(&summary, c::ROOM_CELLS_CULLED));
-        push!(counter_total(&summary, c::ROOM_CELLS_RANGE_CULLED));
-        push!(counter_total(&summary, c::ROOM_VISIBILITY_FALLBACK_DRAWS));
-        push!(counter_total(&summary, c::ROOM_SURF_MATERIAL_CYCLES));
-        push!(counter_total(&summary, c::ROOM_SURF_PROJECTED_CYCLES));
-        push!(counter_total(&summary, c::ROOM_SURF_SCREEN_CYCLES));
-        push!(counter_total(&summary, c::ROOM_SURF_KIND_CYCLES));
-        push!(counter_total(&summary, c::ROOM_SURF_BACKFACE_CYCLES));
-        push!(counter_total(&summary, c::ROOM_SURF_LIGHTING_CYCLES));
-        push!(counter_total(&summary, c::ROOM_SURF_SUBMIT_CYCLES));
-        push!(counter_total(&summary, c::ROOM_SURF_PROFILED));
-        push!(counter_total(&summary, c::ROOM_SURF_SCREEN_CULLED));
-        push!(counter_total(&summary, c::ROOM_SURF_BACKFACE_CULLED));
-        push!(counter_total(&summary, c::ROOM_SUBMIT_HW_SAFE_TEST_CYCLES));
-        push!(counter_total(&summary, c::ROOM_SUBMIT_PACKET_FILL_CYCLES));
-        push!(counter_total(
-            &summary,
-            c::ROOM_SUBMIT_PRIMITIVE_PUSH_CYCLES
-        ));
-        push!(counter_total(&summary, c::ROOM_SUBMIT_DEPTH_CYCLES));
-        push!(counter_total(&summary, c::ROOM_SUBMIT_COMMAND_CYCLES));
-        push!(counter_total(&summary, c::ROOM_SUBMIT_FALLBACK_CYCLES));
-        push!(counter_total(&summary, c::ROOM_SURFACES_CONSIDERED));
-        push!(counter_total(&summary, c::ROOM_PROJECTED_VERTICES));
-        push!(counter_total(&summary, c::ROOM_SURF_WHOLE_QUADS));
-        push!(counter_total(&summary, c::ROOM_SURF_SPLIT_TRIS));
-        push!(counter_total(&summary, c::TRI_PRIMITIVES));
-        push!(counter_latest(c::TRI_PRIMITIVE_REMAINING));
-        push!(counter_total(&summary, c::WORLD_COMMANDS));
-        push!(counter_total(&summary, c::ROOM_SUBMIT_PRIMITIVE_OVERFLOWS));
-        push!(counter_total(&summary, c::MODEL_INSTANCE_DRAWS));
-        push!(counter_total(&summary, c::PLAYER_PROJECTED_VERTICES));
-        push!(counter_total(&summary, c::PLAYER_SUBMITTED_TRIS));
-        push!(counter_total(&summary, c::TEXTURED_MODEL_PARTS));
-        push!(counter_total(&summary, c::TEXTURED_MODEL_VERTICES));
-        push!(counter_total(&summary, c::TEXTURED_MODEL_FACES));
-        push!(counter_total(
-            &summary,
-            c::TEXTURED_MODEL_FAST_SUBMITTED_TRIS
-        ));
-        push!(counter_total(&summary, c::TEXTURED_MODEL_SPLIT_TRIS));
-        push!(counter_total(
-            &summary,
-            c::TEXTURED_MODEL_PACKED_UNCLAMPED_CALLS
-        ));
-        push!(counter_total(
-            &summary,
-            c::TEXTURED_MODEL_PACKED_CLAMPED_CALLS
-        ));
-        push!(counter_total(
-            &summary,
-            c::TEXTURED_MODEL_PACKED_GENERAL_CALLS
-        ));
-        push!(counter_total(
-            &summary,
-            c::TEXTURED_MODEL_FALLBACK_FACE_CALLS
-        ));
-        push!(counter_latest(c::ROOM_STREAM_RESIDENT_MASK_LO));
-        push!(counter_latest(c::ROOM_ACTIVE_CHUNK_MASK_LO));
-        push!(counter_latest(c::ROOM_DRAWN_CHUNK_MASK_LO));
-        push!(counter_latest(c::PORTAL_VIS_VISIBLE_MASK_LO));
-        push!(counter_latest(c::PORTAL_VIS_MISSING_MASK_LO));
-        push!(counter_latest(c::ROOM_CAMERA_GLOBAL_X_BIASED));
-        push!(counter_latest(c::ROOM_CAMERA_GLOBAL_Y_BIASED));
-        push!(counter_latest(c::ROOM_CAMERA_GLOBAL_Z_BIASED));
-        push!(counter_latest(c::ROOM_PLAYER_VIEW_YAW_Q12));
-        push!(counter_latest(c::PORTAL_VIS_CURRENT_ROOM));
-        push!(counter_latest(c::ROOM_PLAYER_ROOM_INDEX));
-
-        debug_assert_eq!(row.len(), PROFILE_LOG_HEADER.len());
-        writeln!(writer, "{}", row.join(",")).map_err(|e| format!("write profile log: {e}"))
-    }
-}
-
 #[cfg(feature = "editor")]
 fn cmd_dump_editor_preview(args: DumpEditorPreviewArgs) -> Result<(), String> {
     let (project_root, project_file) = resolve_project_arg(&args.project);
@@ -2438,321 +1784,12 @@ fn resolve_project_arg(path: &Path) -> (PathBuf, PathBuf) {
     }
 }
 
-fn print_gte_profile(
-    before: &GteProfileSnapshot,
-    after: &GteProfileSnapshot,
-    summary: &telemetry::GuestTelemetrySummary,
-) {
-    let ops = after.ops.saturating_sub(before.ops);
-    let cycles = after
-        .estimated_cycles
-        .saturating_sub(before.estimated_cycles);
-    let guest_frames = summary.frames.max(1);
-    println!("gte_profile:");
-    println!(
-        "  ops={}  per_guest_frame={:.0}",
-        ops,
-        ops as f64 / guest_frames as f64
-    );
-    println!(
-        "  estimated_cycles={}  per_guest_frame={:.0}",
-        cycles,
-        cycles as f64 / guest_frames as f64
-    );
-    // GTE load against the render budget. NOTE: this emulator does not charge
-    // GTE op latency to the CPU frametime, so `estimated_cycles` is the real
-    // hardware GTE load; the rest of the per-frame budget is headroom available
-    // for offloading CPU geometry math (matrix-vector, perspective divide,
-    // cross-product backface) onto the otherwise-idle GTE.
-    let visual_frames = counter_total(summary, telemetry::counter::VISUAL_FRAMES).max(1);
-    let interval_total = counter_total(summary, telemetry::counter::VISUAL_INTERVAL_VBLANKS);
-    let per_render = cycles / visual_frames;
-    if interval_total > 0 && summary.frames > 0 {
-        let vblanks = interval_total as f64 / summary.frames as f64;
-        let budget = vblanks * NTSC_CPU_CYCLES_PER_VBLANK as f64;
-        let pct = if budget > 0.0 {
-            100.0 * per_render as f64 / budget
-        } else {
-            0.0
-        };
-        println!(
-            "  estimated_cycles_per_visual_frame={}  ({:.1}% of the {:.0}-cycle render budget, ~{:.1}% GTE headroom)",
-            per_render,
-            pct,
-            budget,
-            (100.0 - pct).max(0.0)
-        );
-    }
-    println!("  opcodes:");
-    for opcode in 0..after.opcode_counts.len() {
-        let count = after.opcode_counts[opcode].saturating_sub(before.opcode_counts[opcode]);
-        if count == 0 {
-            continue;
-        }
-        println!(
-            "    0x{opcode:02x} {:<6} count={:<10} per_guest_frame={:.0}",
-            gte_opcode_name(opcode as u8),
-            count,
-            count as f64 / guest_frames as f64
-        );
-    }
-}
-
-fn gte_opcode_name(opcode: u8) -> &'static str {
-    match opcode {
-        0x01 => "RTPS",
-        0x06 => "NCLIP",
-        0x0c => "OP",
-        0x10 => "DPCS",
-        0x11 => "INTPL",
-        0x12 => "MVMVA",
-        0x13 => "NCDS",
-        0x14 => "CDP",
-        0x16 => "NCDT",
-        0x1b => "NCCS",
-        0x1c => "CC",
-        0x1e => "NCS",
-        0x20 => "NCT",
-        0x28 => "SQR",
-        0x29 => "DCPL",
-        0x2a => "DPCT",
-        0x2d => "AVSZ3",
-        0x2e => "AVSZ4",
-        0x30 => "RTPT",
-        0x3d => "GPF",
-        0x3e => "GPL",
-        0x3f => "NCCT",
-        _ => "UNKNOWN",
-    }
-}
-
-fn print_guest_profile(summary: &telemetry::GuestTelemetrySummary) {
-    if !summary.has_data() {
-        println!("guest_profile=empty");
-        return;
-    }
-
-    let frames = summary.frames.max(1) as f32;
-    println!("guest_profile_frames={}", summary.frames);
-    println!("guest_profile_frame_meaning=frame_begin_markers");
-    print_guest_pacing_profile(summary);
-    print_guest_render_breakdown(summary);
-    println!("guest_profile_tasks:");
-    for id in 0..telemetry::TASK_COUNT {
-        let cycles = summary.task_cycles[id];
-        if cycles == 0 {
-            continue;
-        }
-        println!(
-            "  {:<18} total={:<10} per_hit={:.0} max_hit={} hits={}",
-            telemetry::task_name(id as u16),
-            cycles,
-            cycles as f32 / (summary.task_hits[id].max(1) as f32),
-            summary.task_max_cycles[id],
-            summary.task_hits[id],
-        );
-    }
-    println!("guest_profile_stages:");
-    for id in 1..telemetry::STAGE_COUNT {
-        let cycles = summary.stage_cycles[id];
-        if cycles == 0 {
-            continue;
-        }
-        println!(
-            "  {:<18} total={:<10} per_frame={:.0} per_hit={:.0} max_hit={} hits={}",
-            telemetry::stage_name(id as u16),
-            cycles,
-            cycles as f32 / frames,
-            cycles as f32 / (summary.stage_hits[id].max(1) as f32),
-            summary.stage_max_cycles[id],
-            summary.stage_hits[id],
-        );
-    }
-    println!("guest_profile_counters:");
-    for id in 1..telemetry::COUNTER_COUNT {
-        let value = summary.counters[id];
-        if value == 0 {
-            continue;
-        }
-        println!(
-            "  {:<18} total={:<10} per_frame={:.0} latest={}",
-            telemetry::counter_name(id as u16),
-            value,
-            value as f32 / frames,
-            summary.counter_latest_values[id],
-        );
-    }
-}
-
-fn print_guest_render_breakdown(summary: &telemetry::GuestTelemetrySummary) {
-    let render_cycles = summary.stage_cycles[telemetry::stage::RENDER as usize];
-    if render_cycles == 0 {
-        println!("guest_profile_render_breakdown=not_emitted");
-        return;
-    }
-
-    let render_hits = summary.stage_hits[telemetry::stage::RENDER as usize].max(1);
-    let mut accounted = 0u64;
-    println!("guest_profile_render_breakdown:");
-    for &(stage_id, label) in GUEST_RENDER_BREAKDOWN_STAGES {
-        let cycles = summary.stage_cycles[stage_id as usize];
-        if cycles == 0 {
-            continue;
-        }
-        accounted = accounted.saturating_add(cycles);
-        println!(
-            "  {:<18} pct={:>5.1} per_render={:.0} cycles={}",
-            label,
-            percent_u64(cycles, render_cycles),
-            cycles as f64 / render_hits as f64,
-            cycles,
-        );
-    }
-
-    let other = render_cycles.saturating_sub(accounted);
-    if other > render_cycles / 200 {
-        println!(
-            "  {:<18} pct={:>5.1} per_render={:.0} cycles={}",
-            "other",
-            percent_u64(other, render_cycles),
-            other as f64 / render_hits as f64,
-            other,
-        );
-    }
-}
-
-fn percent_u64(part: u64, total: u64) -> f64 {
-    (part as f64) * 100.0 / total.max(1) as f64
-}
-
-fn print_guest_pacing_profile(summary: &telemetry::GuestTelemetrySummary) {
-    let pacing_ids = [
-        telemetry::counter::SIM_TICKS,
-        telemetry::counter::VISUAL_FRAMES,
-        telemetry::counter::VISUAL_SKIPPED_VBLANKS,
-        telemetry::counter::VISUAL_DEADLINE_MISSES,
-        telemetry::counter::VISUAL_INTERVAL_VBLANKS,
-        telemetry::counter::VISUAL_MAX_LATENESS_VBLANKS,
-    ];
-    let has_pacing = pacing_ids
-        .iter()
-        .any(|&id| counter_total(summary, id) > 0 || counter_max_value(summary, id) > 0);
-    if !has_pacing {
-        println!("guest_profile_pacing=not_emitted");
-        return;
-    }
-
-    let sim_ticks = counter_total(summary, telemetry::counter::SIM_TICKS);
-    let visual_frames = counter_total(summary, telemetry::counter::VISUAL_FRAMES);
-    let skipped = counter_total(summary, telemetry::counter::VISUAL_SKIPPED_VBLANKS);
-    let misses = counter_total(summary, telemetry::counter::VISUAL_DEADLINE_MISSES);
-    let interval_total = counter_total(summary, telemetry::counter::VISUAL_INTERVAL_VBLANKS);
-    let max_lateness = counter_max_value(summary, telemetry::counter::VISUAL_MAX_LATENESS_VBLANKS);
-    let interval = if summary.frames > 0 && interval_total > 0 {
-        Some(interval_total as f64 / summary.frames as f64)
-    } else {
-        None
-    };
-    let update_per_sim = div_u64(
-        summary.stage_cycles[telemetry::stage::UPDATE as usize],
-        sim_ticks,
-    );
-    let render_per_visual = div_u64(
-        summary.stage_cycles[telemetry::stage::RENDER as usize],
-        visual_frames,
-    );
-    let visual_budget = interval.map(|vblanks| vblanks * NTSC_CPU_CYCLES_PER_VBLANK as f64);
-
-    println!("guest_profile_pacing:");
-    println!("  sim_ticks={}", fmt_known_u64(sim_ticks));
-    println!("  visual_frames={}", fmt_known_u64(visual_frames));
-    println!("  visual_skipped_vblanks={}", skipped);
-    println!("  visual_deadline_misses={}", misses);
-    println!("  visual_interval_vblanks={}", fmt_optional_f64(interval));
-    println!("  visual_max_lateness_vblanks={}", max_lateness);
-    println!(
-        "  update_cycles_per_sim_tick={}",
-        fmt_optional_f64(update_per_sim)
-    );
-    println!(
-        "  render_cycles_per_visual_frame={}",
-        fmt_optional_f64(render_per_visual)
-    );
-    println!(
-        "  visual_budget_cycles={}  vblanks={}  cycles_per_vblank={}",
-        fmt_optional_f64(visual_budget),
-        fmt_optional_f64_2(interval),
-        NTSC_CPU_CYCLES_PER_VBLANK
-    );
-    println!(
-        "  visual_budget_status={}",
-        visual_budget_status(render_per_visual, visual_budget)
-    );
-    println!(
-        "  cadence_status={}",
-        cadence_status(interval, misses, max_lateness)
-    );
-}
-
 fn counter_total(summary: &telemetry::GuestTelemetrySummary, id: u16) -> u64 {
     summary
         .counters
         .get(id as usize)
         .copied()
         .unwrap_or_default()
-}
-
-fn counter_max_value(summary: &telemetry::GuestTelemetrySummary, id: u16) -> u32 {
-    summary
-        .counter_max_values
-        .get(id as usize)
-        .copied()
-        .unwrap_or_default()
-}
-
-fn div_u64(numerator: u64, denominator: u64) -> Option<f64> {
-    (denominator > 0).then_some(numerator as f64 / denominator as f64)
-}
-
-fn fmt_known_u64(value: u64) -> String {
-    if value == 0 {
-        "unknown".to_string()
-    } else {
-        value.to_string()
-    }
-}
-
-fn fmt_optional_f64(value: Option<f64>) -> String {
-    match value {
-        Some(value) => format!("{value:.0}"),
-        None => "unknown".to_string(),
-    }
-}
-
-fn fmt_optional_f64_2(value: Option<f64>) -> String {
-    match value {
-        Some(value) => format!("{value:.2}"),
-        None => "unknown".to_string(),
-    }
-}
-
-fn visual_budget_status(
-    render_per_visual: Option<f64>,
-    visual_budget: Option<f64>,
-) -> &'static str {
-    match (render_per_visual, visual_budget) {
-        (Some(cycles), Some(budget)) if cycles <= budget => "pass",
-        (Some(_), Some(_)) => "fail",
-        _ => "unknown",
-    }
-}
-
-fn cadence_status(interval: Option<f64>, misses: u64, max_lateness: u32) -> &'static str {
-    match interval {
-        Some(_) if misses == 0 && max_lateness == 0 => "steady",
-        Some(_) => "missed_or_late",
-        None => "unknown",
-    }
 }
 
 fn maybe_fast_boot_disc(
